apps/agent-platform/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.25
grpcio==1.59.0
grpcio-tools==1.59.0
pydantic==1.10.14
